内存安全护栏与 CLI 有界化：解码/编码统一闸门，研究命令改为抽样/流式

+++ backend/src/cthulhu_backend/sample_prep/align.py
@@ -7,8 +7,8 @@ import numpy as np
 
 def estimate_shift(reference: np.ndarray, moving: np.ndarray) -> tuple[int, int]:
     """相位相关估计对齐校正量 (dy, dx)：np.roll(moving, (dy, dx)) 后与 reference 重合。"""
-    ref = np.asarray(reference, dtype=np.float64)
-    mov = np.asarray(moving, dtype=np.float64)
+    ref = np.asarray(reference)
+    mov = np.asarray(moving)
     fft_ref = np.fft.fft2(ref)
     fft_mov = np.fft.fft2(mov)
     cross = fft_ref * fft_mov.conj()
@@ -29,8 +29,8 @@ def align_pair(
     border: int = 8,
 ) -> tuple[np.ndarray, np.ndarray, tuple[int, int]]:
     """按平移对齐 moving 并裁出共同区域，返回 (ref_crop, moving_crop, shift)。"""
-    ref = np.asarray(reference, dtype=np.float64)
-    mov = np.asarray(moving, dtype=np.float64)
+    ref = np.asarray(reference)
+    mov = np.asarray(moving)
     dy, dx = shift if shift is not None else estimate_shift(ref, mov)
     rolled = np.roll(np.roll(mov, dy, axis=0), dx, axis=1)
     margin = max(border, abs(dy) + 1, abs(dx) + 1)
@@ -44,8 +44,8 @@ def align_videos(
     moving_frames: np.ndarray,
 ) -> tuple[np.ndarray, np.ndarray, tuple[int, int]]:
     """用中位帧估计平移，逐帧对齐并裁公共区域。"""
-    ref = np.asarray(reference_frames, dtype=np.float64)
-    mov = np.asarray(moving_frames, dtype=np.float64)
+    ref = np.asarray(reference_frames)
+    mov = np.asarray(moving_frames)
     shift = estimate_shift(np.median(ref, axis=0), np.median(mov, axis=0))
     aligned_ref = np.stack([align_pair(ref[i], mov[i], shift)[0] for i in range(min(len(ref), len(mov)))])
     aligned_mov = np.stack([align_pair(ref[i], mov[i], shift)[1] for i in range(min(len(ref), len(mov)))])
